+++ src/test/java/test/TestMoneroCppUtils.java
@@ -9,9 +9,6 @@ import java.util.Map;
 
 import org.junit.Test;
 
-import com.fasterxml.jackson.core.type.TypeReference;
-
-import common.utils.JsonUtils;
 import monero.cpp_bridge.MoneroCppUtils;
 
 /**
@@ -31,7 +28,7 @@ public class TestMoneroCppUtils {
   public void testSerializeHeightsSmall() {
     Map<String, Object> map = new HashMap<String, Object>();
     map.put("heights", Arrays.asList(111, 222, 333));
-    int[] binary = MoneroCppUtils.mapToBinary(map);
+    byte[] binary = MoneroCppUtils.mapToBinary(map);
     assertTrue(binary.length > 0);
     for (int i = 0; i < binary.length; i++) System.out.println(binary[i]);
     Map<String, Object> map2 = MoneroCppUtils.binaryToMap(binary);
@@ -43,7 +40,7 @@ public class TestMoneroCppUtils {
   public void testSerializeHeightsBig() {
     Map<String, Object> map = new HashMap<String, Object>();
     map.put("heights", Arrays.asList(123456, 1234567, 870987));
-    int[] binary = MoneroCppUtils.mapToBinary(map);
+    byte[] binary = MoneroCppUtils.mapToBinary(map);
     assertTrue(binary.length > 0);
     Map<String, Object> map2 = MoneroCppUtils.binaryToMap(binary);
     assertEquals(map, map2);
@@ -54,7 +51,7 @@ public class TestMoneroCppUtils {
   public void testSerializeTextShort() {
     Map<String, Object> map = new HashMap<String, Object>();
     map.put("msg", "Hello there my good man lets make a nice long text to test with lots of exclamation marks!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!");
-    int[] binary = MoneroCppUtils.mapToBinary(map);
+    byte[] binary = MoneroCppUtils.mapToBinary(map);
     assertTrue(binary.length > 0);
     Map<String, Object> map2 = MoneroCppUtils.binaryToMap(binary);
     assertEquals(map, map2);
@@ -78,7 +75,7 @@ public class TestMoneroCppUtils {
             "Hello there my good man lets make a nice long text to test with lots of exclamation marks!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!\n" + 
             "Hello there my good man lets make a nice long text to test with lots of exclamation marks!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!\n" + 
             "Hello there my good man lets make a nice long text to test with lots of exclamation marks!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!\n");
-    int[] binary = MoneroCppUtils.mapToBinary(map);
+    byte[] binary = MoneroCppUtils.mapToBinary(map);
     assertTrue(binary.length > 0);
     Map<String, Object> map2 = MoneroCppUtils.binaryToMap(binary);
     assertEquals(map, map2);
